Change boss health type to Long from Double

+++ src/main/kotlin/click/seichi/gigantic/raid/RaidBoss.kt
@@ -15,20 +15,20 @@ class RaidBoss(boss: Boss) {
 
     val maxHealth = boss.maxHealth
 
-    private val totalDamageMap = mutableMapOf<UUID, Double>()
+    private val totalDamageMap = mutableMapOf<UUID, Long>()
 
     private val dropRelicSet = boss.dropRelicSet
 
-    fun damage(player: Player, amount: Double) {
-        val trueDamage = amount.coerceAtLeast(0.0)
-        val prevTotal = totalDamageMap.getOrDefault(player.uniqueId, 0.0)
+    fun damage(player: Player, amount: Long) {
+        val trueDamage = amount.coerceAtLeast(0L)
+        val prevTotal = totalDamageMap.getOrDefault(player.uniqueId, 0L)
         val nextTotal = prevTotal + trueDamage
         totalDamageMap[player.uniqueId] = nextTotal
         health -= trueDamage
-        health.coerceAtLeast(0.0)
+        health.coerceAtLeast(0L)
     }
 
-    fun getTotalDamage(player: Player) = totalDamageMap[player.uniqueId] ?: 0.0
+    fun getTotalDamage(player: Player) = totalDamageMap[player.uniqueId] ?: 0L
 
     fun getDrop(player: Player): Relic? {
         if (!isDefeated(player)) return null
@@ -41,6 +41,6 @@ class RaidBoss(boss: Boss) {
         return totalDamage >= maxHealth.div(10.0).times(1.0)
     }
 
-    fun isDead() = health <= 0.0
+    fun isDead() = health <= 0L
 
 }
